move ListEcrViewer
